# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VirtuProfs
basePath: /api
host: localhost:5100
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  passport:
    type: apiKey
    in: header
    name: Authorization
  admin:
    type: apiKey
    in: header
    name: Authorization
  teacher:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /subscriptions:
    x-swagger-router-controller: SubscriptionController
    get:
      operationId: getSubscriptions
      summary: Get all subscriptions
      tags:
        - subscription
      security:
        - passport: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /subscriptions/active:
    x-swagger-router-controller: SubscriptionController
    get:
      operationId: activeSubscriptions
      summary: Get active subscriptions
      tags:
        - subscription
      security:
        - passport: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /subscriptions/ending:
    x-swagger-router-controller: BalanceController
    get:
      operationId: findEnding
      summary: Get all subscriptions
      tags:
        - subscription
      security:
        - passport: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
  /subscriptions/manage:
    x-swagger-router-controller: SubscriptionController
    get:
      operationId: getSubscriptionsManage
      summary: Get all subscriptions
      tags:
        - subscription
      security:
        - admin: []
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    post:
      operationId: setSubscription
      summary: Set subscription
      tags:
        - subscription
      security:
        - admin: []
      parameters:
        - in: body
          name: subscription
          description: Subscription set
          schema:
            type: object
            properties:
              type:
                type: number
                minimum: 0
                maximum: 2
              duration:
                type: number
              validTo:
                type: string
              status:
                type: boolean
              repeatable:
                type: boolean
              currency:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
  /subscription:
    x-swagger-router-controller: SubscriptionController
    get:
      operationId: findSubscription
      summary: find subscription
      tags:
        - subscription
      security:
        - passport: []
      parameters:
        - name: code
          in: query
          type: string
          description: find subscription
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    put:
      operationId: updateSubscription
      summary: Update subscription
      tags:
        - subscription
      security:
        - admin: []
      parameters:
        - in: body
          name: subscription
          description: Subscription set
          schema:
            type: object
            properties:
              _id:
                type: string
              user:
                type: string
              type:
                type: number
                minimum: 0
                maximum: 2
              duration:
                type: number
              validTo:
                type: string
              status:
                type: number
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    post:
      operationId: buySubject
      summary: buy Subject
      tags:
        - subscription
      security:
        - passport: []
      parameters:
        - in: body
          name: subscription
          description: buy Subject
          schema:
            type: object
            required:
              - subject
              - period
            properties:
              subject:
                type: string
              period:
                type: string
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    delete:
      operationId: deleteSubscription
      summary: Delete subscription
      security:
        - admin: []
      tags:
        - subscription
      parameters:
        - name: ids
          in: query
          type: array
          collectionFormat: csv
          items:
            type: string
          description: Delete subscriptions by ID
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
  /settings:
    x-swagger-router-controller: SettingsController
    put:
      operationId: setSettings
      summary: Update settings
      security:
        - passport: []
        - teacher: []
        - admin: []
      tags:
        - settings
      parameters:
        - in: body
          name: settings
          description: Settings object
          schema:
            $ref: "#/definitions/Settings"
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
    get:
      operationId: getSettings
      summary: Get all settings
      tags:
        - settings
      responses:
        "200":
          description: Successful response
          schema:
            $ref: "#/definitions/Response"
...
definitions:
  Response:
    type: object
    properties:
      success:
        type: boolean
      errorCode:
        type: string
      errorStack:
        type: array
        items:
            type: object
      message:
        type: string
      data:
        type: object
    required:
      - success
...
  Settings:
    type: object
    properties:
      name:
        type: string
      value:
        type: number
      _id:
        type: string